requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.5
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
